backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.4
python-docx==1.1.2
docxtpl==0.16.8
Pillow==11.1.0
openai==1.59.6
python-multipart==0.0.20
